// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      comments: {
        Row: {
          blog_post_id: string | null
          content: string
          created_at: string
          id: string
          journal_id: string | null
          parent_comment_id: string | null
          user_id: string
        }
        Insert: {
          blog_post_id?: string | null
          content: string
          created_at?: string
          id?: string
          journal_id?: string | null
          parent_comment_id?: string | null
          user_id: string
        }
        Update: {
          blog_post_id?: string | null
          content?: string
          created_at?: string
          id?: string
          journal_id?: string | null
          parent_comment_id?: string | null
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "comments_blog_post_id_fkey"
            columns: ["blog_post_id"]
            isOneToOne: false
            referencedRelation: "blog_posts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "comments_journal_id_fkey"
            columns: ["journal_id"]
            isOneToOne: false
            referencedRelation: "journals"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "comments_parent_comment_id_fkey"
            columns: ["parent_comment_id"]
            isOneToOne: false
            referencedRelation: "comments"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}